refactor(broker): introduce PartitionManager interface
